# client-api/swagger.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: Aprendendo Swagger
  description: Fazendo Curso Udemy. [Saiba Mais](https://www.udemy.com/microservices-rest-com-swagger-na-pratica)
  
consumes: 
  - application/json
produces: 
  - application/json

tags:
  - name: 'Consulta'
  - name: 'Cadastro'
  - name: 'Manutencao'
  - name: 'Exclusao'
  - name: 'API'

#localhost:8080/cliente-api/v1

host: localhost:8080
basePath: /cliente-api/v1
schemes: 
  - http
  - https

paths:
  /clientes:
    get:
      tags:
        - Consulta
      summary: Todos os clientes
      operationId: consultaTodosClientes
      responses:
        200:
          description: "Sucesso"
          schema:
            $ref: '#/definitions/Clientes'
        400:
          description: "Consulta inválida"
        401:
          description: "Consulta não autorizada"
        404:
          description: "Nenhum cliente encontrado"
        500:
          description: "Erro interno do servidor"

  /cliente:
    post:
      tags:
        - Cadastro
      summary: Grava um cliente
      operationId: cadastraCliente
      description: Grava um cliente
      parameters: 
        - name: "cliente"
          in: "body"
          required: true
          schema:
            $ref: '#/definitions/Cliente'
      responses:
        201:
          description: "Cadastrado"
          schema:
            $ref: '#/definitions/Cliente'
          headers:
            Location:
              type: "string"
              description: "URI do cliente alterado"
        400:
          description: "Cadastro inválido"
        401:
          description: "Cadastro não autorizada"
        500:
          description: "Erro interno do servidor no módulo de cadastro de cliente"
  
  /cliente/{id}:
    put:
      tags: 
        - Manutencao
      summary: "Altera um cliente existente"
      operationId: alteraCliente
      parameters: 
        - in: "path"
          type: "integer"
          format: "int32"
          name: "id"
          required: true
        - in: "body"
          name: "cliente"
          required: true
          schema:
            $ref: '#/definitions/Cliente'
      responses:
        202:
          description: "Sucesso na alteração"
          schema:
            $ref: '#/definitions/Cliente'
          headers:
            Location:
              type: "string"
              description: "URI do cliente alterado"
        400:
          description: "Edit inválido"
        401:
          description: "Edit não autorizado"
        500:
          description: "Erro interno do servidor no módulo de alteração de cliente"
        
    get:
      tags: 
        - Consulta
      summary: "Consulta um cliente pelo id"
      operationId: consultaClientePorId
      parameters: 
        - in: "path"
          type: "integer"
          format: "int32"
          name: "id"
          required: true
        
      responses:
        200:
          description: "Sucesso na consulta"
          schema:
            $ref: '#/definitions/Cliente'
          headers:
            Location:
              type: "string"
              description: "URI do cliente alterado"
        400:
          description: "Consulta inválido"
        401:
          description: "Consulta não autorizada"
        404:
          description: "Não encontrou nenhum cliente"
        500:
          description: "Erro interno do servidor no módulo de consulta de cliente por id"      
    
    delete:
      tags: 
        - Exclusao
      summary: "Apaga um cliente pelo id"
      operationId: apagaClientePorId
      parameters: 
        - in: "path"
          type: "integer"
          format: "int32"
          name: "id"
          required: true
        
      responses:
        204:
          description: "Cliente apagado com sucesso"
        400:
          description: "Consulta inválido"
        401:
          description: "Consulta não autorizada"
        404:
          description: "Não encontrou nenhum cliente"
        500:
          description: "Erro interno do servidor no módulo de consulta de cliente por id"    
  
  /cliente/consulta/{sobrenome}:
    get:
      tags: 
        - Consulta
      summary: "Consulta um cliente pelo sobrenome"
      operationId: consultaSobrenome
      parameters: 
        - in: "path"
          type: "string"
          name: "sobrenome"
          required: true
      responses:
        200:
          description: "Sucesso na consulta por sobrenome"
          schema:
            $ref: '#/definitions/Clientes'
          headers:
            Location:
              type: "string"
              description: "URI do cliente alterado"
        400:
          description: "Consulta inválida"
        401:
          description: "Consulta não autorizado"
        404:
          description: "Nada encontrado"
        500:
          description: "Erro interno do servidor no módulo de consulta por sobrenome" 
          
  /cliente/{status}/{id}:
    put:
      tags: 
        - Manutencao
      summary: "Altera o status pelo id do cliente"
      operationId: alterarStatusById
      parameters: 
        - in: "path"
          type: "string"
          name: "status"
          enum: ["ativo", "inativo"]
          required: true
        - in: "path"
          name: "id"
          type: "integer"
          format: "int32"
          required: true
      responses:
        202:
          description: "Sucesso ao alterar status"
          schema:
            $ref: '#/definitions/Cliente'
          headers:
            Location:
              type: "string"
              description: "URI do cliente alterado"
        400:
          description: "Consulta inválida"
        401:
          description: "Alteração não autorizada"
        404:
          description: "Nada encontrado, por isso não houve alteração"
        500:
          description: "Erro interno do servidor no módulo de alteração de status" 
  /readme:
    get:
      tags: 
        - API
      summary: "Recupera readme da API"
      operationId: readme
      responses:
        200:
          description: "Sucesso na leitura do readme"
          schema:
            type: string
        500:
          description: "Erro interno do servidor no módulo de consulta API" 

definitions:
  Clientes:
    type: "array"
    items:
      $ref: '#/definitions/Cliente'
    
  Cliente:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int32"
      nome:
        type: "string"
      sobrenome:
        type: "string"
      iniciais:
        type: "string"
        minLength: 2
        maxLength: 7
      dataNascimento:
        type: "string"
        format: "date"
      dataAtualizacaoRegistro:
        type: "string"
        format: "date-time"
      status:
        type: "boolean"
        description: "Ativo ou Inativo"
      tipo:
        type: "string"
        enum: ["Interno", "Externo"]
  
        
    